backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose
passlib[bcrypt]
bcrypt
pydantic[email]
email-validator
python-dotenv
pytest
httpx
ics
celery
redis
slowapi
pyinstrument
pillow
